almost done

+++ post/views.py
@@ -1,7 +1,9 @@
-from django.shortcuts import render, HttpResponse, get_object_or_404, HttpResponseRedirect, redirect
+from django.shortcuts import render, HttpResponse, get_object_or_404, HttpResponseRedirect, redirect, Http404
 from .models import Post
 from .forms import PostForm
 from django.contrib import messages
+from django.utils.text import slugify
+
 
 def post_index(request):
     posts = Post.objects.all()
@@ -33,10 +35,16 @@ def post_create(request):
     #     # Formu kullanıcıya göster
     #     form = PostForm()
 
+    if not request.user.is_authenticated():
+        # Eğer kullanıcı giriş yapmamış ise hata sayfası gönder
+        return Http404()
+
     form = PostForm(request.POST or None, request.FILES or None)
     if form.is_valid():
-        post = form.save()
-        messages.success(request, "Basarili bir sekilde olusturdunuz.")
+        post = form.save(commit=False)
+        post.slug = slugify(post.title)
+        post.save()
+        messages.success(request, "Başarılı bir şekilde oluşturdunuz.", extra_tags='mesaj-basarili')
         return HttpResponseRedirect(post.get_absolute_url())
 
     context = {
@@ -47,11 +55,16 @@ def post_create(request):
 
 
 def post_update(request, id):
+
+    if not request.user.is_authenticated():
+        # Eğer kullanıcı giriş yapmamış ise hata sayfası gönder
+        return Http404()
+
     post = get_object_or_404(Post, id=id)
-    form = PostForm(request.POST or None,request.FILES or None, instance=post)
+    form = PostForm(request.POST or None, request.FILES or None, instance=post)
     if form.is_valid():
         form.save()
-        messages.success(request, "Basarili bir sekilde guncellediniz.", extra_tags='mesaj-basarili')
+        messages.success(request, "Başarılı bir şekilde güncellediniz.")
         return HttpResponseRedirect(post.get_absolute_url())
 
     context = {
@@ -62,6 +75,11 @@ def post_update(request, id):
 
 
 def post_delete(request, id):
+
+    if not request.user.is_authenticated():
+        # Eğer kullanıcı giriş yapmamış ise hata sayfası gönder
+        return Http404()
+
     post = get_object_or_404(Post, id=id)
     post.delete()
     return redirect("post:index")
